requirements.txt
streamlit==1.28.0
langchain==0.0.350
chromadb==0.4.15
sentence-transformers==2.2.2
pypdf2==3.0.1
python-dotenv==1.0.0
groq==0.32.0
huggingface-hub==0.17.0
python-docx==0.8.11
faiss-cpu==1.7.4
numpy==1.24.4
torch==2.0.1
transformers==4.35.2
requests==2.31.0
urllib3==1.26.18